templates/flask/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-Cors==4.0.1